Add tests for launching learn more link
Also modifies the existing test for the SubscribeNow page's
launcher to unify the launcher mocks.
