+++ gui/packages/ubuntupro/test/pages/landscape_skip/landscape_page_test.dart
@@ -2,13 +2,18 @@ import 'package:flutter/material.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:ubuntupro/pages/landscape_skip/landscape_skip_page.dart';
+import 'package:url_launcher_platform_interface/url_launcher_platform_interface.dart';
 import 'package:wizard_router/wizard_router.dart';
 import 'package:yaru/yaru.dart';
 import 'package:yaru_test/yaru_test.dart';
 
 import '../../utils/build_multiprovider_app.dart';
+import '../../utils/url_launcher_mock.dart';
 
 void main() {
+  final launcher = FakeUrlLauncher();
+  UrlLauncherPlatform.instance = launcher;
+
   testWidgets('default state', (tester) async {
     final app = buildApp();
     await tester.pumpWidget(app);
@@ -80,6 +85,18 @@ void main() {
       tester.widget<YaruSelectableContainer>(skipRadioTile).selected,
       isTrue,
     );
+  });
+
+  testWidgets('launch web page', (tester) async {
+    final app = buildApp();
+    await tester.pumpWidget(app);
+    final context = tester.element(find.byType(LandscapeSkipPage));
+    final lang = AppLocalizations.of(context);
+
+    expect(launcher.launched, isFalse);
+    await tester.tapOnText(find.textRange.ofSubstring(lang.learnMore));
+    await tester.pump();
+    expect(launcher.launched, isTrue);
   });
 }
