+++ gui/packages/ubuntupro/test/utils/url_launcher_mock.dart
@@ -1,0 +1,48 @@
+import 'package:url_launcher_platform_interface/link.dart';
+import 'package:url_launcher_platform_interface/url_launcher_platform_interface.dart';
+
+class FakeUrlLauncher extends UrlLauncherPlatform {
+  bool launched = false;
+
+  @override
+  Future<bool> canLaunch(String url) async {
+    return true;
+  }
+
+  @override
+  Future<void> closeWebView() async {}
+
+  @override
+  Future<bool> launchUrl(String url, LaunchOptions options) async {
+    launched = true;
+    return true;
+  }
+
+  @override
+  Future<bool> supportsCloseForMode(PreferredLaunchMode mode) async {
+    return true;
+  }
+
+  @override
+  Future<bool> supportsMode(PreferredLaunchMode mode) async {
+    return true;
+  }
+
+  @override
+  Future<bool> launch(
+    String url, {
+    required bool useSafariVC,
+    required bool useWebView,
+    required bool enableJavaScript,
+    required bool enableDomStorage,
+    required bool universalLinksOnly,
+    required Map<String, String> headers,
+    String? webOnlyWindowName,
+  }) async {
+    launched = true;
+    return true;
+  }
+
+  @override
+  LinkDelegate? get linkDelegate => null;
+}
